Module-7/requirements.txt
groq
python-dotenv
faiss-cpu
sentence-transformers
numpy
streamlit